requirements.txt
Flask==1.1.1
line-bot-sdk==1.16.0
pya3rt
requests==2.12.4
beautifulsoup4==4.7.1
urllib3==1.24.1
soupsieve==1.6.1